boostcamp/requirements.txt
## The following requirements were added by pip freeze:
alembic==1.11.1
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
databricks-cli==0.17.7
docker==6.1.3
entrypoints==0.4
Flask==2.3.2
gitdb==4.0.10
GitPython==3.1.31
gunicorn==20.1.0
idna==3.4
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
mlflow==1.24.0
numpy==1.20.0
oauthlib==3.2.2
packaging==23.1
pandas==1.4.4
prometheus-client==0.17.0
prometheus-flask-exporter==0.22.4
protobuf==3.20.1
PyJWT==2.7.0
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
querystring-parser==1.2.4
requests==2.31.0
scipy==1.10.1
six==1.16.0
smmap==5.0.0
SQLAlchemy==2.0.18
sqlparse==0.4.4
tabulate==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
websocket-client==1.6.1
Werkzeug==2.3.6
zipp==3.15.0